improve typing, place version check to utils, rename sanic workers env variable
