+++ rasa_sdk/constants.py
@@ -1,3 +1,3 @@
 DEFAULT_SERVER_PORT = 5055
 DEFAULT_SANIC_WORKERS = 1
-ENV_SANIC_WORKERS = "SANIC_WORKERS"
+ENV_SANIC_WORKERS = "ACTION_SERVER_SANIC_WORKERS"
